requirements.txt
certifi==2020.12.5
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.13
django-appconf==1.0.4
django-compressor==2.4
django-cors-headers==3.6.0
django-libsass==0.8
django-stdimage==5.1.1
django-widget-tweaks==1.4.2
djangorestframework==3.12.2
googlemaps==4.4.2
gunicorn==19.7.1
idna==2.6
instamojo-wrapper==1.1
libsass==0.20.1
pi==0.1.2
Pillow==8.1.0
pipenv==11.10.1
progressbar2==3.53.1
psycopg2==2.8.4
PyJWT==1.7.1
python-utils==2.5.5
pytz==2018.4
rcssmin==1.0.6
requests==2.20.0
rjsmin==1.1.0
sentry-sdk==0.19.5
six==1.15.0
sqlparse==0.4.1
twilio==6.51.0
urllib3==1.24.3
uWSGI==2.0.17
virtualenv==15.2.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
